requirements-gpu.txt
# Requirements for GPU version (torch installed separately in Dockerfile)
numpy<2
python-telegram-bot==21.3
asyncpg==0.29.0
pgvector==0.3.2
sqlalchemy[asyncio]==2.0.31
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
Pillow==10.4.0
httpx==0.27.0
folium==0.15.1
staticmap==0.5.7
transformers==4.38.0
open-clip-torch==2.24.0
rembg==2.0.57
onnxruntime-gpu==1.17.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
